# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/internal/wallets" : {
      "post" : {
        "tags" : [ "internal-wallet-controller" ],
        "operationId" : "createWallet",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateWalletRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WalletDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/wallets/{entityType}/{entityId}/debit" : {
      "post" : {
        "tags" : [ "internal-wallet-controller" ],
        "operationId" : "debit",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "ADVERTISER" ]
          }
        }, {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TransactionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WalletDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/wallets/{entityType}/{entityId}/credit" : {
      "post" : {
        "tags" : [ "internal-wallet-controller" ],
        "operationId" : "credit",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "ADVERTISER" ]
          }
        }, {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TransactionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WalletDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/advertisers/{advertiserId}/wallet/topups" : {
      "post" : {
        "tags" : [ "wallet-topup-controller" ],
        "operationId" : "topupWallet",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TopupWalletRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMapStringString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/wallet/dlq/retry" : {
      "post" : {
        "tags" : [ "admin-dlq-controller" ],
        "operationId" : "retryDlqEvent",
        "parameters" : [ {
          "name" : "topic",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "object"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/wallet/dlq/outbox/{eventId}/retry" : {
      "post" : {
        "tags" : [ "admin-dlq-controller" ],
        "operationId" : "retryOutboxDlqEvent",
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/money/advertiser/{entityType}/{entityId}" : {
      "get" : {
        "tags" : [ "payee-wallet-controller" ],
        "operationId" : "getWallet",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "ADVERTISER" ]
          }
        }, {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WalletDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/money/advertiser/{entityType}/{entityId}/transactions" : {
      "get" : {
        "tags" : [ "payee-wallet-controller" ],
        "operationId" : "getTransactions",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "ADVERTISER" ]
          }
        }, {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoWalletTransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/money/advertiser/{entityType}/{entityId}/topups/{topupId}" : {
      "get" : {
        "tags" : [ "payee-wallet-controller" ],
        "operationId" : "getTopupStatus",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "ADVERTISER" ]
          }
        }, {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "topupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/wallets/{entityType}/{entityId}" : {
      "get" : {
        "tags" : [ "internal-wallet-controller" ],
        "operationId" : "getWallet_1",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "ADVERTISER" ]
          }
        }, {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WalletDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/wallets/{entityType}/{entityId}/transactions" : {
      "get" : {
        "tags" : [ "internal-wallet-controller" ],
        "operationId" : "getTransactions_1",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "ADVERTISER" ]
          }
        }, {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoWalletTransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/wallets/transactions/reference/{referenceId}" : {
      "get" : {
        "tags" : [ "internal-wallet-controller" ],
        "operationId" : "getTransactionByReference",
        "parameters" : [ {
          "name" : "referenceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WalletTransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/wallets/balances" : {
      "get" : {
        "tags" : [ "internal-wallet-controller" ],
        "operationId" : "getBalances",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoWalletDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/wallet/dlq/outbox" : {
      "get" : {
        "tags" : [ "admin-dlq-controller" ],
        "operationId" : "getOutboxDlqEvents",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoOutboxEventEntity"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "CreateWalletRequest" : {
        "required" : [ "currency" ],
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "entityType" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "ADVERTISER" ]
          },
          "currency" : {
            "type" : "string"
          }
        }
      },
      "WalletDto" : {
        "required" : [ "balance", "currency", "entityId", "entityType", "id", "status" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "entityId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "entityType" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "ADVERTISER" ]
          },
          "balance" : {
            "type" : "number"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "SUSPENDED", "CLOSED" ]
          }
        }
      },
      "TransactionRequest" : {
        "required" : [ "amount", "category", "referenceId" ],
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "referenceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "category" : {
            "type" : "string",
            "enum" : [ "DELIVERY_FEE", "PLATFORM_FIXED_FEE", "PLATFORM_BONUS", "FOOD_COST", "SGST", "CGST", "REFUND", "ORDER_TOTAL", "AD_IMPRESSION", "AD_CLICK", "AD_CONVERSION", "AD_WALLET_TOPUP", "CLAWBACK", "PAYOUT_TRANSFER", "CASH_COLLECTED", "CASH_REMITTED", "STORE_CREDIT" ]
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "TopupWalletRequest" : {
        "required" : [ "amount" ],
        "type" : "object",
        "properties" : {
          "amount" : {
            "minimum" : 0.01,
            "exclusiveMinimum" : false,
            "type" : "number"
          },
          "gatewayName" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseMapStringString" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseString" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageResponseDtoWalletTransactionDto" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WalletTransactionDto"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "WalletTransactionDto" : {
        "required" : [ "amount", "createdAt", "id", "transactionType", "walletId" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "walletId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "type" : "number"
          },
          "transactionType" : {
            "type" : "string",
            "enum" : [ "CREDIT", "DEBIT" ]
          },
          "referenceId" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "metadata" : {
            "type" : "string"
          }
        }
      },
      "PageResponseDtoWalletDto" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WalletDto"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "OutboxEventEntity" : {
        "required" : [ "aggregateId", "aggregateType", "createdAt", "eventType", "id", "payload", "retryCount", "status" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "aggregateType" : {
            "type" : "string",
            "enum" : [ "ORDER", "PAYMENT", "NOTIFICATION", "OUTLET", "BRAND", "LEDGER", "ADVERTISEMENT", "WALLET", "CHAT_SESSION", "REVIEW" ]
          },
          "aggregateId" : {
            "type" : "string"
          },
          "eventType" : {
            "type" : "string",
            "enum" : [ "ORDER_CREATED", "ORDER_PAID", "ORDER_PLACED_COD", "ORDER_ACCEPTED", "ORDER_PREPARING", "ORDER_READY", "ORDER_DELIVERED", "ORDER_REJECTED", "ORDER_AT_RESTAURANT", "ORDER_STATUS_UPDATED", "ORDER_STATUS_SYNC", "ORDER_CANCELLED", "ORDER_CANCELLED_BY_RESTAURANT", "ORDER_CANCELLED_BY_CUSTOMER", "ORDER_CANCELLED_BY_ADMIN", "ORDER_DELAY_APPROVAL_REQUESTED", "ORDER_DELAY_APPROVED", "ORDER_DELAY_REJECTED", "DISPATCH_CANDIDATE_FOUND", "DISPATCH_FAILED", "DRIVER_ASSIGNED", "ORDER_DRIVER_REJECTED", "MANUAL_INTERVENTION_REQUIRED", "FORCE_ASSIGN_DRIVER", "DELIVERY_FAILED", "NOTIFICATION_REQUEST", "NOTIFICATION_DISPATCH", "PAYMENT_WEBHOOK", "PAYMENT_COMPLETED", "PAYMENT_FAILED", "PAYMENT_REFUNDED", "PAYMENT_REFUND_REQUESTED", "PAYMENT_REFUND_FAILED", "REFUND_REQUESTED", "REFUND_FAILED", "PAYMENT_PARTIALLY_REFUNDED", "ORDER_PARTIALLY_REFUNDED", "LEDGER_TRANSACTION_REQUEST", "OUTLET_ACTIVATED", "MENU_UPDATED", "OUTLET_DEACTIVATED", "BRAND_CREATED", "AD_CAMPAIGN_CREATED", "AD_CAMPAIGN_UPDATED", "AD_CAMPAIGN_PAUSED", "AD_CAMPAIGN_RESUMED", "AD_CAMPAIGN_COMPLETED", "AD_CAMPAIGN_DELETED", "AD_CREATIVE_PENDING", "AD_CREATIVE_APPROVED", "AD_CREATIVE_REJECTED", "AD_CAMPAIGN_BUDGET_EXHAUSTED", "AD_CAMPAIGN_PACING_UPDATED", "AD_IMPRESSION_BILLED", "AD_CLICK_BILLED", "AD_CONVERSION_BILLED", "AD_WALLET_TOPUP_REQUEST", "AD_WALLET_TOPUP_COMPLETED", "AD_BUDGET_ALERT", "CHAT_REFUND_QUOTE_REQUESTED", "CHAT_REFUND_REQUESTED", "CHAT_REFUND_QUOTE_RESPONSE", "CHAT_REFUND_DECISION", "CHAT_REFUND_ERROR", "REVIEW_CREATED" ]
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "payload" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "UNPROCESSED", "PROCESSED", "FAILED", "DLQ" ]
          },
          "processedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "retryCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "new" : {
            "type" : "boolean"
          }
        }
      },
      "PageResponseDtoOutboxEventEntity" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OutboxEventEntity"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      }
    }
  }
}